Make the SQL shim pluggable.  Move SQLClient to sql_client.cc. Add example sql_shim plugin.
