+++ include/mysql/plugin_auth.h.pp
@@ -53,6 +53,92 @@ struct st_mysql_plugin
   void * __reserved1;
   unsigned long flags;
 };
+#include "com_data.h"
+typedef struct st_com_init_db_data
+{
+  const char *db_name;
+  unsigned long length;
+} COM_INIT_DB_DATA;
+typedef struct st_com_refresh_data
+{
+  unsigned char options;
+} COM_REFRESH_DATA;
+typedef struct st_com_shutdown_data
+{
+  enum mysql_enum_shutdown_level level;
+} COM_SHUTDOWN_DATA;
+typedef struct st_com_kill_data
+{
+  unsigned long id;
+} COM_KILL_DATA;
+typedef struct st_com_set_option_data
+{
+  unsigned int opt_command;
+} COM_SET_OPTION_DATA;
+typedef struct st_com_stmt_execute_data
+{
+  unsigned long stmt_id;
+  unsigned long flags;
+  unsigned char *params;
+  unsigned long params_length;
+} COM_STMT_EXECUTE_DATA;
+typedef struct st_com_stmt_fetch_data
+{
+  unsigned long stmt_id;
+  unsigned long num_rows;
+} COM_STMT_FETCH_DATA;
+typedef struct st_com_stmt_send_long_data_data
+{
+  unsigned long stmt_id;
+  unsigned int param_number;
+  unsigned char *longdata;
+  unsigned long length;
+} COM_STMT_SEND_LONG_DATA_DATA;
+typedef struct st_com_stmt_prepare_data
+{
+  const char *query;
+  unsigned int length;
+} COM_STMT_PREPARE_DATA;
+typedef struct st_stmt_close_data
+{
+  unsigned int stmt_id;
+} COM_STMT_CLOSE_DATA;
+typedef struct st_com_stmt_reset_data
+{
+  unsigned int stmt_id;
+} COM_STMT_RESET_DATA;
+typedef struct st_com_query_data
+{
+  const char *query;
+  unsigned int length;
+} COM_QUERY_DATA;
+typedef struct st_com_field_list_data
+{
+  unsigned char *table_name;
+  unsigned int table_name_length;
+  const unsigned char *query;
+  unsigned int query_length;
+} COM_FIELD_LIST_DATA;
+union COM_DATA {
+  COM_INIT_DB_DATA com_init_db;
+  COM_REFRESH_DATA com_refresh;
+  COM_SHUTDOWN_DATA com_shutdown;
+  COM_KILL_DATA com_kill;
+  COM_SET_OPTION_DATA com_set_option;
+  COM_STMT_EXECUTE_DATA com_stmt_execute;
+  COM_STMT_FETCH_DATA com_stmt_fetch;
+  COM_STMT_SEND_LONG_DATA_DATA com_stmt_send_long_data;
+  COM_STMT_PREPARE_DATA com_stmt_prepare;
+  COM_STMT_CLOSE_DATA com_stmt_close;
+  COM_STMT_RESET_DATA com_stmt_reset;
+  COM_QUERY_DATA com_query;
+  COM_FIELD_LIST_DATA com_field_list;
+};
+struct st_mysql_sqlshim
+{
+  int interface_version;
+  my_bool (*shim_function)(void*, union COM_DATA*, enum enum_server_command, union COM_DATA*, enum enum_server_command*);
+};
 struct st_mysql_daemon
 {
   int interface_version;
