+++ sql/sql_plugin.cc
@@ -102,7 +102,8 @@ const LEX_STRING plugin_type_names[MYSQL_MAX_PLUGIN_TYPE_NUM]=
   { C_STRING_WITH_LEN("AUTHENTICATION") },
   { C_STRING_WITH_LEN("VALIDATE PASSWORD") },
   { C_STRING_WITH_LEN("GROUP REPLICATION") },
-  { C_STRING_WITH_LEN("KEYRING") }
+  { C_STRING_WITH_LEN("KEYRING") },
+  { C_STRING_WITH_LEN("SQL SHIM") }
 };
 
 extern int initialize_schema_table(st_plugin_int *plugin);
@@ -157,7 +158,8 @@ static int min_plugin_info_interface_version[MYSQL_MAX_PLUGIN_TYPE_NUM]=
   MYSQL_AUTHENTICATION_INTERFACE_VERSION,
   MYSQL_VALIDATE_PASSWORD_INTERFACE_VERSION,
   MYSQL_GROUP_REPLICATION_INTERFACE_VERSION,
-  MYSQL_KEYRING_INTERFACE_VERSION
+  MYSQL_KEYRING_INTERFACE_VERSION,
+  MYSQL_SQLSHIM_INTERFACE_VERSION
 };
 static int cur_plugin_info_interface_version[MYSQL_MAX_PLUGIN_TYPE_NUM]=
 {
@@ -171,7 +173,8 @@ static int cur_plugin_info_interface_version[MYSQL_MAX_PLUGIN_TYPE_NUM]=
   MYSQL_AUTHENTICATION_INTERFACE_VERSION,
   MYSQL_VALIDATE_PASSWORD_INTERFACE_VERSION,
   MYSQL_GROUP_REPLICATION_INTERFACE_VERSION,
-  MYSQL_KEYRING_INTERFACE_VERSION
+  MYSQL_KEYRING_INTERFACE_VERSION,
+  MYSQL_SQLSHIM_INTERFACE_VERSION
 };
 
 /* support for Services */
@@ -736,8 +739,7 @@ static void plugin_dl_del(const LEX_STRING *dl)
 #endif
 }
 
-
-static st_plugin_int *plugin_find_internal(const LEX_CSTRING &name,
+st_plugin_int *plugin_find_internal(const LEX_CSTRING &name,
                                                   int type)
 {
   uint i;
@@ -764,6 +766,11 @@ static st_plugin_int *plugin_find_internal(const LEX_CSTRING &name,
                        reinterpret_cast<const uchar*>(name.str),
                        name.length));
   DBUG_RETURN(NULL);
+}
+
+extern "C" plugin_ref plugin_get_sql_shim() {
+    LEX_CSTRING plugin_name = { C_STRING_WITH_LEN("sql_shim") };
+    return plugin_find_internal(plugin_name, MYSQL_SQLSHIM_PLUGIN);
 }
